move js/interop tests to paredit-config
