+++ src/extension-test/unit/cursor-doc/paredit-config-test.ts
@@ -3,7 +3,6 @@ import * as paredit from '../../../cursor-doc/paredit';
 import * as pareditConfig from '../../../cursor-doc/paredit-config';
 import * as model from '../../../cursor-doc/model';
 import { docFromTextNotation, getText, textAndSelection } from '../common/text-notation';
-import { defaultBindingForms } from '../../../cursor-doc/paredit-config';
 
 model.initScanner(20000);
 
@@ -354,6 +353,60 @@ describe('paredit-config', () => {
         const config = pareditConfig.createPareditConfig(customForms, {}, { m: 'my.ns' });
         paredit.growSelection(a, a.selections, config);
         expect(getText(a)).toBe(getText(b));
+      });
+    });
+
+    describe('js-interop/let vector-binding form', () => {
+      it('applied-science.js-interop/let - grows selection to binding pairs', () => {
+        const a = docFromTextNotation('(applied-science.js-interop/let [a b |c| d])');
+        const aSelection = a.selections[0];
+        const b = docFromTextNotation('(applied-science.js-interop/let [a b |c d|])');
+        const bSelection = b.selections[0];
+        paredit.growSelection(a);
+        expect(a.selectionsStack).toEqual([[aSelection], [bSelection]]);
+      });
+    });
+
+    describe('js-interop flat pair forms', () => {
+      it('applied-science.js-interop/assoc! - grows selection to key-value pairs', () => {
+        const a = docFromTextNotation('(applied-science.js-interop/assoc! obj :a 1 |:b| 2)');
+        const aSelection = a.selections[0];
+        const b = docFromTextNotation('(applied-science.js-interop/assoc! obj :a 1 |:b 2|)');
+        const bSelection = b.selections[0];
+        paredit.growSelection(a);
+        expect(a.selectionsStack).toEqual([[aSelection], [bSelection]]);
+      });
+
+      it('applied-science.js-interop/obj - grows selection to key-value pairs starting at offset 1', () => {
+        const a = docFromTextNotation('(applied-science.js-interop/obj |:a| 1 :b 2)');
+        const aSelection = a.selections[0];
+        const b = docFromTextNotation('(applied-science.js-interop/obj |:a 1| :b 2)');
+        const bSelection = b.selections[0];
+        paredit.growSelection(a);
+        expect(a.selectionsStack).toEqual([[aSelection], [bSelection]]);
+      });
+
+      it('applied-science.js-interop/assoc! - does not treat object argument as part of pair', () => {
+        const a = docFromTextNotation('(applied-science.js-interop/assoc! |obj| :a 1 :b 2)');
+        const aSelection = a.selections[0];
+        const b = docFromTextNotation('(|applied-science.js-interop/assoc! obj :a 1 :b 2|)');
+        const bSelection = b.selections[0];
+        paredit.growSelection(a);
+        expect(a.selectionsStack).toEqual([[aSelection], [bSelection]]);
+      });
+
+      it('j/assoc! - works with alias map configuration', () => {
+        const a = docFromTextNotation('(j/assoc! obj :a 1 |:b| 2)');
+        const aSelection = a.selections[0];
+        const b = docFromTextNotation('(j/assoc! obj :a 1 |:b 2|)');
+        const bSelection = b.selections[0];
+        const config = pareditConfig.createPareditConfig(
+          [],
+          {},
+          { j: 'applied-science.js-interop' }
+        );
+        paredit.growSelection(a, a.selections, config);
+        expect(a.selectionsStack).toEqual([[aSelection], [bSelection]]);
       });
     });
